+++ headers/entities/pawns/CoopBot.h
@@ -6,7 +6,7 @@ class BulletPool;
 
 class CoopBot final : public Bot
 {
-	void TickUpdate(float deltaTime) override;
+	void TickUpdate(double deltaTime) override;
 
 public:
 	CoopBot(PawnProperty pawnProperty, const std::shared_ptr<BulletPool>& bulletPool, BonusEffectProperty effects = {},
