+++ source/components/managers/AnimationManager.cpp
@@ -43,7 +43,7 @@ void AnimationManager::Subscribe()
 	_events->AddListener("GameModeChangedTo", _name, [this](const GameMode newGameMode) { SetGameMode(newGameMode); });
 	_events->AddListener("PostDraw", _name, [this]() { Update(); });
 	_events->AddListener("AnimationTankUpdate", _name, [this](const std::string& objName) { UpdateTank(objName); });
-	_events->AddListener("PostTickUpdate", _name, [this](const float /*deltaTime*/) { this->AnimationSeqDisposer(); });
+	_events->AddListener("PostTickUpdate", _name, [this](const double /*deltaTime*/) { this->AnimationSeqDisposer(); });
 }
 
 void AnimationManager::SubscribeAsHost()
